Refactor ProductItem for reusability
